# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almha API",
        "description": "API para el sistema administrativo de Almha Plastic Surgery",
        "contact": {
            "email": "admin@almhaplasticsurgery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/ping": {
            "get": {
                "summary": "Prueba de Swagger",
                "operationId": "20ad7da9e80f9197be89d18be4dc869b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Servidor de API v1"
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "28b6b601899c3d9607e41f71f106605e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refrescar token",
                "operationId": "269a401d46b02b7446d38bf9883d3083",
                "responses": {
                    "200": {
                        "description": "Token refrescado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Listar todos los miembros del equipo",
                "operationId": "ccb74d663ac6f50c60c757a38bb587d6",
                "responses": {
                    "200": {
                        "description": "Lista de miembros del equipo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Crear un nuevo miembro del equipo",
                "operationId": "dc3e02724aa7a96daee1831eb2c64c6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "userId",
                                    "name",
                                    "specialty",
                                    "description"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Imagen de perfil",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "userId": {
                                        "description": "UUID del usuario vinculado",
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "specialty": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "description": "Galería de imágenes",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Miembro creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/{id}": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Obtener detalle de un miembro del equipo",
                "operationId": "bc9a4ecf8922392e97fff99f120c7600",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Actualizar un miembro del equipo",
                "operationId": "0010e232d0d20a6afa7791924f1a2420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "baseLang",
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "baseLang": {
                                        "type": "string",
                                        "example": "es"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "userId": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Nueva imagen de perfil",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "biography": {
                                        "type": "string"
                                    },
                                    "gallery": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "path": {
                                                    "type": "string",
                                                    "format": "binary"
                                                },
                                                "order": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Miembro actualizado exitosamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Team"
                ],
                "summary": "Eliminar un miembro del equipo",
                "operationId": "d75be9480a1fb7ff8ba36bd1ef29e2cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miembro eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "d012eebed7ed10747f238a3669338532",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "d480868e336a24f1717cc1978b75748f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "7aea9d42f31ac055ef37b244536a0bc4",
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Team",
            "description": "Team"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}